Move auth logic to AspNetCore package
