+++ src/FunderMaps.WebApi/Controllers/Application/OrganizationUserAdminController.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
-using FunderMaps.AspNetCore.Authentication;
 using FunderMaps.AspNetCore.DataTransferObjects;
+using FunderMaps.AspNetCore.Services;
 using FunderMaps.Core.Entities;
 using FunderMaps.Core.Exceptions;
 using FunderMaps.Core.Interfaces.Repositories;
@@ -30,7 +30,7 @@ namespace FunderMaps.WebApi.Controllers.Application
         private readonly IMapper _mapper;
         private readonly IUserRepository _userRepository;
         private readonly IOrganizationUserRepository _organizationUserRepository;
-        private readonly SignInService _signinService;
+        private readonly SignInService _signInService;
 
         /// <summary>
         ///     Create new instance.
@@ -39,12 +39,12 @@ namespace FunderMaps.WebApi.Controllers.Application
             IMapper mapper,
             IUserRepository userRepository,
             IOrganizationUserRepository organizationUserRepository,
-            SignInService signinService)
+            SignInService signInService)
         {
             _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
             _userRepository = userRepository ?? throw new ArgumentNullException(nameof(userRepository));
             _organizationUserRepository = organizationUserRepository ?? throw new ArgumentNullException(nameof(organizationUserRepository));
-            _signinService = signinService ?? throw new ArgumentNullException(nameof(signinService));
+            _signInService = signInService ?? throw new ArgumentNullException(nameof(signInService));
         }
 
         // POST: api/admin/organization/{id}/user
@@ -60,7 +60,7 @@ namespace FunderMaps.WebApi.Controllers.Application
             // Act.
             // FUTURE: Do in 1 call.
             user = await _userRepository.AddGetAsync(user);
-            await _signinService.SetPasswordAsync(user.Id, input.Password);
+            await _signInService.SetPasswordAsync(user.Id, input.Password);
             await _organizationUserRepository.AddAsync(id, user.Id, input.OrganizationRole);
 
             // Map.
@@ -150,7 +150,7 @@ namespace FunderMaps.WebApi.Controllers.Application
             }
 
             // Act.
-            await _signinService.SetPasswordAsync(userId, input.NewPassword);
+            await _signInService.SetPasswordAsync(userId, input.NewPassword);
 
             // Return.
             return NoContent();
